perf: preload encounters on hover
